requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
click==8.1.3
coverage==6.4.4
Deprecated==1.2.13
Django==4.1.1
django-dotenv==1.4.2
django-rest-framework==0.1.0
django-rq==2.5.1
djangorestframework==3.13.1
flake8==5.0.4
mccabe==0.7.0
packaging==21.3
psycopg2==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
rq==1.11.0
sqlparse==0.4.2
wrapt==1.14.1
